# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "description": "Returns every user in the system with public fields only: id, email, fullName, numeric role, createdAt, updatedAt, lastLoginAt. Passwords and refresh token hashes are never returned.",
        "operationId": "UsersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "ServiceResultContainer<UserResponseDto[]>.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersListResultResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "500": {
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all users",
        "summaryHe": "מציגים את רשימת המשתמשים הפעילים במערכת",
        "tags": [
          "users"
        ],
        "toolIcon": "ph-users"
      }
    },
    "/users/me": {
      "get": {
        "description": "Returns the JWT payload decoded by JwtAuthGuard. No database query is made. Role is numeric: 1 = Admin, 2 = User.",
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "ServiceResultContainer<JwtPayload> with sub, email, numeric role, iat, and exp.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JwtPayloadResultResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "500": {
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current authenticated user payload",
        "summaryHe": "מציגים את פרטי הפרופיל המהירים של המשתמש הנוכחי",
        "tags": [
          "users"
        ],
        "toolIcon": "ph-user-circle"
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Fetches one user by numeric id with public fields only. Use GET /users first to discover valid IDs.",
        "operationId": "UsersController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique numeric user id. ParseIntPipe rejects non-numeric values.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ServiceResultContainer<UserResponseDto>.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResultResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "404": {
            "description": "No user found with the given id."
          },
          "500": {
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by id",
        "summaryHe": "מציגים פרטים מלאים על משתמש לפי מזהה ייחודי",
        "tags": [
          "users"
        ],
        "toolIcon": "ph-user"
      },
      "patch": {
        "description": "Updates fullName and/or email for the user identified by :id. This endpoint intentionally does not accept role changes; use PATCH /users/:id/role for role updates.",
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric id of the user to update.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Partial profile update. Example: { \"fullName\": \"New Name\" } or { \"email\": \"new@example.com\" }.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ServiceResultContainer<UserResponseDto> with the updated user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResultResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "403": {
            "description": "Authenticated user is not an admin."
          },
          "404": {
            "description": "No user found with the given id."
          },
          "500": {
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user profile fields",
        "summaryHe": "מעדכנים את פרטי הפרופיל האישיים של המשתמש",
        "tags": [
          "users"
        ],
        "toolIcon": "ph-pencil-simple"
      },
      "delete": {
        "description": "Hard-deletes the user record. This is irreversible and requires admin privileges.",
        "operationId": "UsersController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric id of the user to delete.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion confirmed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserResultResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "403": {
            "description": "Authenticated user is not an admin."
          },
          "404": {
            "description": "No user found with the given id."
          },
          "500": {
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete user permanently",
        "summaryHe": "מוחקים משתמש לצמיתות מהמערכת",
        "tags": [
          "users"
        ],
        "toolIcon": "ph-trash"
      }
    },
    "/users/{id}/role": {
      "patch": {
        "description": "Sets only the role field of the user identified by :id. IMPORTANT: role must be a NUMBER: 1 = Admin, 2 = User. Never send a string. Accepted numeric values : 1 = Admin, 2 = User. Other user fields must be updated through PATCH /users/:id.",
        "operationId": "UsersController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric id of the user whose role will change.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Role payload. Example: { \"role\": 1 } for Admin or { \"role\": 2 } for User.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ServiceResultContainer<UserResponseDto> with the updated user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResultResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "403": {
            "description": "Authenticated user is not an admin."
          },
          "404": {
            "description": "No user found with the given id."
          },
          "500": {
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change user role",
        "summaryHe": "מעדכנים את תפקיד והרשאות המשתמש במערכת",
        "tags": [
          "users"
        ],
        "toolIcon": "ph-shield"
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UserRole": {
        "type": "number",
        "enum": [
          1,
          2
        ],
        "description": "Numeric user role. 1 = Admin, 2 = User."
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique numeric user id.",
            "example": 1
          },
          "email": {
            "type": "string",
            "description": "User email address.",
            "example": "user@example.com"
          },
          "fullName": {
            "type": "string",
            "description": "Full display name of the user.",
            "example": "John Doe"
          },
          "role": {
            "description": "Numeric user role. 1 = Admin, 2 = User.",
            "example": 2,
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the user was created.",
            "example": "2026-05-12T10:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the user was last updated.",
            "example": "2026-05-12T11:00:00Z"
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the user last successful login, or null if the user never logged in.",
            "example": "2026-05-20T10:00:00Z",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "role",
          "createdAt",
          "updatedAt",
          "lastLoginAt"
        ]
      },
      "UserResultResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request succeeded.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable result message.",
            "example": "User retrieved successfully"
          },
          "result": {
            "description": "Public user response payload.",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserResponseDto"
              }
            ]
          }
        },
        "required": [
          "success",
          "message",
          "result"
        ]
      },
...
      "JwtPayloadResponseDto": {
        "type": "object",
        "properties": {
          "sub": {
            "type": "number",
            "description": "Authenticated user id from the JWT subject.",
            "example": 1
          },
          "email": {
            "type": "string",
            "description": "Authenticated user email from the JWT payload.",
            "example": "user@example.com"
          },
          "role": {
            "description": "Numeric user role from the JWT payload. 1 = Admin, 2 = User.",
            "example": 2,
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "iat": {
            "type": "number",
            "description": "JWT issued-at timestamp in seconds.",
            "example": 1770000000
          },
          "exp": {
            "type": "number",
            "description": "JWT expiry timestamp in seconds.",
            "example": 1770000900
          }
        },
        "required": [
          "sub",
          "email",
          "role",
          "iat",
          "exp"
        ]
      },
      "JwtPayloadResultResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request succeeded.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable result message.",
            "example": "Current user retrieved successfully"
          },
          "result": {
            "description": "Decoded JWT payload.",
            "allOf": [
              {
                "$ref": "#/components/schemas/JwtPayloadResponseDto"
              }
            ]
          }
        },
        "required": [
          "success",
          "message",
          "result"
        ]
      },
      "UsersListResultResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request succeeded.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable result message.",
            "example": "Users retrieved successfully"
          },
          "result": {
            "description": "Public user response list.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponseDto"
            }
          }
        },
        "required": [
          "success",
          "message",
          "result"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "John Updated",
            "description": "Full display name of the user. Optional - omit this field to leave the current value unchanged. Validation: if provided, must be a string and must not be empty."
          },
          "email": {
            "type": "string",
            "example": "updated@example.com",
            "format": "email",
            "description": "New email address for the user. Optional - omit to keep the current email. Validation: must be a valid email format if provided. Uniqueness is enforced by the service/database layer."
          }
        }
      },
      "DeleteUserPayloadDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "description": "True when the user was deleted.",
            "example": true
          }
        },
        "required": [
          "deleted"
        ]
      },
      "DeleteUserResultResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request succeeded.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable result message.",
            "example": "User deleted successfully"
          },
          "result": {
            "description": "Deletion result payload.",
            "allOf": [
              {
                "$ref": "#/components/schemas/DeleteUserPayloadDto"
              }
            ]
          }
        },
        "required": [
          "success",
          "message",
          "result"
        ]
      },
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "example": 2,
            "description": "The new numeric role to assign to the user. Required - this DTO represents a role-only update. Accepted numeric values: 1 = Admin, 2 = User. Security note: use 1 with caution because it grants access to admin-only endpoints.",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          }
        },
        "required": [
          "role"
        ]
      },
...
    }
  }
}